requirements.txt
# Core ML libraries
transformers==4.41.0
peft==0.11.1
accelerator==0.29.2
bitsandbytes==0.43.0
bitsandbytes-cuda122==0.43.0

# Data processing
datasets==2.19.0
pandas==2.0.3
numpy==1.24.3
tqdm==4.66.1
faiss-gpu==1.7.4

# Recommender systems
recsim-ng==1.1.3
sasrec-pytorch==0.1.2

# NLP
sentencepiece==0.1.99

# Experiment tracking
wandb==0.17.0

# Evaluation
perspective-api-client==1.1.0

# Utilities
pyyaml==6.0.1
dm-tree==0.1.8
